+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - A new play mode for groups, Random, which will trigger a random child on play
 - A counter showing selected items in the footer
 - A shortcut for aborting an item silently (Shift + F1)
+- References and triggers can now target the main client's selection
 ### Changed
 - Added 'Stop' as an option when setting default on end actions in settings
 - Electron now bypasses proxies on requests to localhost